wip effing up router
