+++ src/hooks/use-routes-service.js
@@ -1,0 +1,8 @@
+import { useContext } from "React"
+import { RoutesServiceContext } from "../routes/app"
+
+export function useRoutesService() {
+  let routesService = useContext(RoutesServiceContext)
+
+  return routesService
+}
